refactor: reconstruct UserName component in accordance with atomic design
