+++ src/components/molecules/user/UserNameIdentified.tsx
@@ -1,0 +1,9 @@
+import BaseUserName from './BaseUserName';
+
+type Props = {
+  userName: string;
+};
+
+export default function UserNameIdentified({ userName }: Props) {
+  return <BaseUserName className="text-primary">{userName}</BaseUserName>;
+}
